Digitizeink/backend/requirements.txt

flask
fastapi
uvicorn
pytesseract
easyocr
opencv-python
pillow
numpy
pymongo
werkzeug
flask-cors
